requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
discord==2.0.0
discord.py==2.0.1
ffxivweather==1.0.14
frozenlist==1.3.1
idna==3.4
jsonpickle==2.2.0
multidict==6.0.2
mysql-connector-python==8.0.31
protobuf==3.20.1
psycopg2-binary==2.9.4
python-dotenv==0.21.0
pytz==2022.5
yarl==1.8.1
